+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Upgraded Dear ImGui version used in example to [v1.89.9 WIP](https://github.com/ocornut/imgui/commit/11613013860d149667302a258041dcd832069f36)
 - Added CMake support, see CMakeLists.txt for details
 - Optional dismiss button for notifications added
+- Added an optional button to the notification that executes a user-defined function
 
 ## [0.0.1] - 28.07.2023